Move `getApi` to new `api` subdirectory.
