+++ frontend/src/app/functions/api/index.ts
@@ -1,0 +1,1 @@
+export { getApi } from './getApi';
